+++ CHANGELOG.md
@@ -1,6 +1,35 @@
 Changelog
 ===
 
+## Version 9.15.1
+### 🛠 Fixes & Updates
+
+* more callout icon tweaks ([#1137](https://github.com/readmeio/markdown/issues/1137)) ([ab658c5](https://github.com/readmeio/markdown/commit/ab658c5fc2d5c380cc7d252fedc9e516cb745796))
+
+## Version 9.15.0
+### ✨ New & Improved
+
+* jsx mermaid ([#1133](https://github.com/readmeio/markdown/issues/1133)) ([5f22e49](https://github.com/readmeio/markdown/commit/5f22e49dccdd692588338685489a61da1dc7749a))
+
+### 🛠 Fixes & Updates
+
+* **components:** add PostmanRunButton ([#1136](https://github.com/readmeio/markdown/issues/1136)) ([618687f](https://github.com/readmeio/markdown/commit/618687fbd7a0e7a746cd5799e09b4fe2416fb602))
+
+## Version 9.14.0
+### ✨ New & Improved
+
+* code text wrap ([#1131](https://github.com/readmeio/markdown/issues/1131)) ([cf9fe06](https://github.com/readmeio/markdown/commit/cf9fe067b5416dc370ea9a0b660cf77e4045076c))
+
+### 🛠 Fixes & Updates
+
+* callout emoji w/o h3 ([#1132](https://github.com/readmeio/markdown/issues/1132)) ([f9e3dfd](https://github.com/readmeio/markdown/commit/f9e3dfda718ee9c5ede03af27119ec09d0fa7a4d))
+* fa icon styling ([#1130](https://github.com/readmeio/markdown/issues/1130)) ([564d242](https://github.com/readmeio/markdown/commit/564d24278a2ca66f81a19c514f890e2d77e6a0b2))
+
+## Version 9.13.0
+### ✨ New & Improved
+
+* better callout defaults ([#1129](https://github.com/readmeio/markdown/issues/1129)) ([bbdc057](https://github.com/readmeio/markdown/commit/bbdc057eba5a74ca1bd3c083cba48eb89a09abb8))
+
 ## Version 9.12.0
 ### ✨ New & Improved